refactor : 기능에 맞게 키워드 통합 테스트 분리
